+++ src/Symfony/Bundle/FrameworkBundle/Command/ContainerLintCommand.php
@@ -77,7 +77,7 @@ final class ContainerLintCommand extends Command
         $kernel = $this->getApplication()->getKernel();
         $kernelContainer = $kernel->getContainer();
 
-        if (!$kernel->isDebug() || !(new ConfigCache($kernelContainer->getParameter('debug.container.dump'), true))->isFresh()) {
+        if (!$kernel->isDebug() || !$kernelContainer->getParameter('debug.container.dump') || !(new ConfigCache($kernelContainer->getParameter('debug.container.dump'), true))->isFresh()) {
             if (!$kernel instanceof Kernel) {
                 throw new RuntimeException(sprintf('This command does not support the application kernel: "%s" does not extend "%s".', get_debug_type($kernel), Kernel::class));
             }